scripts/requirements.txt
singlestoredb==1.4.0
python-dotenv==1.0.0
